hadoop-hdfs: Modify FileIoProvider interface to support zonefs
To abstract zonefs unique file operation requirements, the new class
ZoneFsFileIoProvider will be introduced as a sub-class of
FileIoProvider. To prepare for this ZoneFsFileIoProvider, add new
methods to FileIoProvider: getPosition(), truncate() and length(). Also
modify arguments of transferToSocketFully() and syncFileRange(). Also
modify the modifier of a few fields and methods from private to
protected to allow reference from sub-class.

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>